pyproject.toml
[project]
name = "kaggle408"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.0.0",
    "dotenv>=0.9.9",
    "jupyter>=1.1.1",
    "openai>=1.97.0",
    "pillow>=11.3.0",
]
